deformetricar: an R client for the Deformetrica shape-registration toolkit

deformetricar wraps the Deformetrica (>= 4.3) command-line tools for statistical shape analysis: it fits diffeomorphisms between point clouds, neuron backbones and surface meshes, and applies them by geodesic shooting. It also reads and writes VTK object formats.

Package options

deformetricar.exe

Path to (or name of) the deformetrica executable; consulted by find_deformetrica. Set it in your Rprofile if Deformetrica is not on the PATH or in a conda environment.

References

https://www.deformetrica.org

See also

Author

Maintainer: Alexander Bates alexander.shakeel.bates@gmail.com

Authors:

Gregory Jefferis jefferis@gmail.com (ORCID)